Cleanup some code reduplication for sonarcloud
