+++ front-end/src/app/layout/sidebar/menus/f24/f24-menu.component.ts
@@ -1,103 +1,83 @@
 import { Component, OnInit } from '@angular/core';
-import { combineLatest, Observable, of, switchMap, takeUntil } from 'rxjs';
 import { Store } from '@ngrx/store';
 import { MenuItem } from 'primeng/api';
-import { selectActiveReport } from '../../../../store/active-report.selectors';
 import { Report } from '../../../../shared/models/report.model';
 import { ReportService } from '../../../../shared/services/report.service';
 import { ReportSidebarSection, SidebarState } from '../../sidebar.component';
-import { selectSidebarState } from 'app/store/sidebar-state.selectors';
-import { DestroyerComponent } from 'app/shared/components/app-destroyer.component';
+import { AbstractMenuComponent } from '../abstract-menu.component';
 
 @Component({
   selector: 'app-f24-menu',
   templateUrl: './f24-menu.component.html',
   styleUrls: ['../menu-report.component.scss'],
 })
-export class F24MenuComponent extends DestroyerComponent implements OnInit {
-  activeReport$?: Observable<Report | undefined>;
-  items$: Observable<MenuItem[]> = of([]);
-
+export class F24MenuComponent extends AbstractMenuComponent implements OnInit {
   formLabel?: string;
 
-  constructor(private store: Store, private reportService: ReportService) {
-    super();
+  constructor(store: Store, reportService: ReportService) {
+    super(store, reportService);
   }
 
-  ngOnInit(): void {
-    this.activeReport$ = this.store.select(selectActiveReport);
-
-    this.activeReport$.pipe(takeUntil(this.destroy$)).subscribe((report) => {
-      this.formLabel = report?.formLabel;
-    });
-
-    this.items$ = combineLatest([this.store.select(selectSidebarState), this.activeReport$]).pipe(
-      takeUntil(this.destroy$),
-      switchMap(([sidebarState, activeReport]: [SidebarState, Report | undefined]) => {
-        const isEditable = this.reportService.isEditable(activeReport);
-
-        const transactionItems = [
+  getMenuItems(sidebarState: SidebarState, activeReport: Report | undefined, isEditable: boolean): MenuItem[] {
+    const transactionItems = [
+      {
+        label: 'Manage your transactions',
+        routerLink: `/reports/transactions/report/${activeReport?.id}/list`,
+      },
+      {
+        label: 'Add an independent expenditure',
+        routerLink: `/reports/f24/report/${activeReport?.id}/transactions/select/independent-expenditures`,
+      },
+    ];
+    return [
+      {
+        label: 'ENTER A TRANSACTION',
+        expanded: sidebarState?.section == ReportSidebarSection.TRANSACTIONS,
+        visible: isEditable,
+        items: transactionItems,
+      },
+      {
+        label: 'REVIEW TRANSACTIONS',
+        expanded: sidebarState?.section == ReportSidebarSection.TRANSACTIONS,
+        visible: !isEditable,
+        routerLink: `/reports/transactions/report/${activeReport?.id}/list`,
+      },
+      {
+        label: 'REVIEW A REPORT',
+        expanded: sidebarState?.section == ReportSidebarSection.REVIEW,
+        items: [
           {
-            label: 'Manage your transactions',
-            routerLink: `/reports/transactions/report/${activeReport?.id}/list`,
+            label: 'View print preview',
+            routerLink: `/reports/f24/web-print/${activeReport?.id}`,
           },
           {
-            label: 'Add an independent expenditure',
-            routerLink: `/reports/f24/report/${activeReport?.id}/transactions/select/independent-expenditures`,
-          },
-        ];
-
-        return of([
-          {
-            label: 'ENTER A TRANSACTION',
-            expanded: sidebarState?.section == ReportSidebarSection.TRANSACTIONS,
+            label: 'Add a report level memo',
+            routerLink: `/reports/f24/memo/${activeReport?.id}`,
             visible: isEditable,
-            items: transactionItems,
+          },
+        ],
+      },
+      {
+        label: 'SIGN & SUBMIT',
+        expanded: sidebarState?.section == ReportSidebarSection.SUBMISSION,
+        items: [
+          {
+            label: 'Confirm information',
+            routerLink: `/reports/f24/submit/step1/${activeReport?.id}`,
+            visible: isEditable,
           },
           {
-            label: 'REVIEW TRANSACTIONS',
-            expanded: sidebarState?.section == ReportSidebarSection.TRANSACTIONS,
+            label: 'Submit report',
+            routerLink: `/reports/f24/submit/step2/${activeReport?.id}`,
+            visible: isEditable,
+          },
+          {
+            label: 'Report status',
+            routerLink: `/reports/f24/submit/status/${activeReport?.id}`,
             visible: !isEditable,
-            routerLink: `/reports/transactions/report/${activeReport?.id}/list`,
           },
-          {
-            label: 'REVIEW A REPORT',
-            expanded: sidebarState?.section == ReportSidebarSection.REVIEW,
-            items: [
-              {
-                label: 'View print preview',
-                routerLink: `/reports/f24/web-print/${activeReport?.id}`,
-              },
-              {
-                label: 'Add a report level memo',
-                routerLink: `/reports/f24/memo/${activeReport?.id}`,
-                visible: isEditable,
-              },
-            ],
-          },
-          {
-            label: 'SIGN & SUBMIT',
-            expanded: sidebarState?.section == ReportSidebarSection.SUBMISSION,
-            items: [
-              {
-                label: 'Confirm information',
-                routerLink: `/reports/f24/submit/step1/${activeReport?.id}`,
-                visible: isEditable,
-              },
-              {
-                label: 'Submit report',
-                routerLink: `/reports/f24/submit/step2/${activeReport?.id}`,
-                visible: isEditable,
-              },
-              {
-                label: 'Report status',
-                routerLink: `/reports/f24/submit/status/${activeReport?.id}`,
-                visible: !isEditable,
-              },
-            ],
-          },
-        ] as MenuItem[]);
-      })
-    );
+        ],
+      },
+    ] as MenuItem[];
   }
 }
